+++ cmake/TestseT.cmake
@@ -54,7 +54,6 @@ add_eT_runtest(hf_energy_cartesian_basis_set                "eT;short;hf;sad;car
 add_eT_runtest(hf_energy_external_basis                     "eT;short;hf;sad;cartesian;externalbasis")
 add_eT_runtest(hf_geoopt                                    "eT;short;hf;bfgs;gradient")
 add_eT_runtest(hf_scf_energy                                "eT;short;hf;sad")
-add_eT_runtest(hf_mo_energy                                 "eT;short;hf;mo-diis")
 add_eT_runtest(hf_energy_qmmmnopol                          "eT;short;hf;sad;qmmm;mmnopol")
 add_eT_runtest(hf_energy_qmfq                               "eT;short;hf;sad;qmmm;fq")
 add_eT_runtest(hf_scf_energy_qmmmnopol                      "eT;short;hf;sad;qmmm;mmnopol")