+++ src/wavefunctions/mp2/mean_value_mp2.F90
@@ -17,10 +17,10 @@
 !  along with this program. If not, see <https://www.gnu.org/licenses/>.
 !
 !
-submodule (mp2_class) zop_mp2
+submodule (mp2_class) mean_value_mp2
 !
 !!
-!!    Zeroth order properties submodule 
+!!    Mean-value submodule 
 !!
 !!    Contains routines related to the mean values, i.e. 
 !!    the construction of density matrices as well as expectation 
@@ -99,4 +99,4 @@ contains
    end subroutine calculate_energy_mp2
 !
 !
-end submodule zop_mp2
+end submodule mean_value_mp2