+++ changelog.md
@@ -26,6 +26,7 @@
 - Parallel excited states are now removed from the wavefunction if found eT-program/eT:!717
 - Added keyword to control storage in micro iterations of Newton-Raphson ground state solver (`micro iterations storage`). See merge request eT-program/eT:!651
 - Non-linear davidson: minimal threshold in the microiterations set to half the residual threshold. See merge request eT-program/eT!738
+- The conversion script removes `scf_restart_file` and `cc_restart_file` after it has been used. See merge request eT-program/eT!745
 
 ### Structure
 - Omega and Jacobian routines have separate in and out vectors and call the corresponding parent routines. See merge request eT-program/eT:!684
@@ -36,7 +37,13 @@
 - A single SCF solver now handles all SCF calculations, it can be used with different types of convergence acceleration (none, DIIS, CROP). See merge request eT-program/eT:!654
 - Removed some unused routines from reordering and array utilities. See merge request eT-program/eT:!725
 - Newton-Raphson solver for coupled cluster ground state is now handled by the same solver as the default DIIS solver. See merge request eT-program/eT:!651
-+ Optimized the autogenerate script and added functionality to generate complex modules from real modules. See merge request eT-program/eT:!735
+- Optimized the autogenerate script and added functionality to generate complex modules from real modules. See merge request eT-program/eT:!735
+
+
+# eT v1.1.3
+### Structure
+- Made Python script to update all test refences and updated the files. See merge request eT-program/eT:!746
+
 
 # eT v1.1.2
 ### Bugfixes